✅ KZGZ77 code: done
Task: 202604260814-KZGZ77
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: Split runner type contracts into focused modules under runner/types while preserving the runner/types.js facade | details: typecheck, runner tests, knip, formatting, bootstrap, and doctor passed.
