+++ packages/agentplane/src/runner/types/constants.ts
@@ -1,0 +1,3 @@
+export const RUNNER_BUNDLE_SCHEMA_VERSION = 1 as const;
+export const RUNNER_API_VERSION = "1" as const;
+export const RUNNER_TRACE_SCHEMA_VERSION = 1 as const;
