+++ packages/agentplane/src/runner/types/state.ts
@@ -1,0 +1,107 @@
+import type { RunnerTimeoutReason } from "@agentplaneorg/core/config";
+import type { AgentplaneCapabilityRegistry } from "../../runtime/capabilities/index.js";
+
+import type { RunnerAdapterCapabilities } from "./capabilities.js";
+import type { RunnerContextBundle, RunnerExecutionContract } from "./context.js";
+import type { RUNNER_API_VERSION, RUNNER_BUNDLE_SCHEMA_VERSION } from "./constants.js";
+import type {
+  RunnerInvocation,
+  RunnerResult,
+  RunnerResultStatus,
+  RunnerTracePolicy,
+  RunnerTimeoutPolicy,
+} from "./invocation.js";
+import type { RunnerPolicyDecision } from "./policy.js";
+import type { RunnerTarget } from "./target.js";
+
+export type RunnerProcessSignal = "SIGHUP" | "SIGINT" | "SIGQUIT" | "SIGTERM" | "SIGKILL";
+
+export type RunnerSupervisionState = {
+  pid?: number | null;
+  command?: string | null;
+  started_at?: string | null;
+  heartbeat_at?: string | null;
+  cancel_requested_at?: string | null;
+  cancel_signal?: RunnerProcessSignal | null;
+  timeout_reason?: RunnerTimeoutReason | null;
+  timeout_requested_at?: string | null;
+  terminate_sent_at?: string | null;
+  kill_sent_at?: string | null;
+  force_killed?: boolean;
+  exit_signal?: RunnerProcessSignal | null;
+};
+
+export type RunnerLifecycleStatus = "prepared" | "running" | RunnerResultStatus;
+
+export type RunnerInvocationSnapshot = {
+  executable: string | null;
+  argv: string[];
+  argv_count: number;
+  env_keys: string[];
+  cwd: string | null;
+  run_dir: string | null;
+  bundle_path: string | null;
+  state_path: string | null;
+  events_path: string | null;
+  result_path: string | null;
+  trace_path: string | null;
+  stderr_path: string | null;
+  bootstrap_path: string | null;
+  output_last_message_path: string | null;
+  dry_run: boolean;
+  has_result_path: boolean;
+  has_output_last_message_path: boolean;
+};
+
+export type RunnerPreparedMetadata = {
+  prompt_count: number;
+  bundle_bytes: number;
+  bootstrap_bytes: number;
+  bundle_sha256: string;
+  bootstrap_sha256: string;
+  has_task_context: boolean;
+  has_recipe_context: boolean;
+  trace_policy: RunnerTracePolicy;
+  timeout_policy: RunnerTimeoutPolicy;
+  adapter_capabilities?: RunnerAdapterCapabilities;
+  adapter_capability_registry?: AgentplaneCapabilityRegistry;
+  policy_decision?: RunnerPolicyDecision;
+  invocation: RunnerInvocationSnapshot;
+};
+
+export type RunnerRunState = {
+  schema_version: typeof RUNNER_BUNDLE_SCHEMA_VERSION;
+  runner_api_version: typeof RUNNER_API_VERSION;
+  run_id: string;
+  adapter_id: string;
+  target: RunnerTarget;
+  status: RunnerLifecycleStatus;
+  mode: RunnerExecutionContract["mode"];
+  bundle_path: string;
+  result_path: string;
+  bootstrap_path?: string | null;
+  events_path: string;
+  trace_path: string;
+  stderr_path: string;
+  trace_policy: RunnerTracePolicy;
+  timeout_policy: RunnerTimeoutPolicy;
+  policy_decision?: RunnerPolicyDecision;
+  created_at: string;
+  updated_at: string;
+  prepared_metadata?: RunnerPreparedMetadata;
+  supervision?: RunnerSupervisionState;
+  result?: RunnerResult;
+};
+
+export type RunnerEvent = {
+  at: string;
+  type: string;
+  message: string;
+  data?: Record<string, unknown>;
+};
+
+export type RunnerRunRecord = {
+  bundle: RunnerContextBundle;
+  state: RunnerRunState;
+  result: RunnerResult | null;
+};
