+++ packages/agentplane/src/runner/types/target.ts
@@ -1,0 +1,11 @@
+export type RunnerTarget =
+  | {
+      kind: "task";
+      task_id: string;
+    }
+  | {
+      kind: "recipe_scenario";
+      recipe_id: string;
+      scenario_id: string;
+      task_id?: string;
+    };
